@@ -24,12 +24,13 @@
 
 #pragma once
 
- 
+
 #include "util/NumType.h"
 #include "util/IndexThreadReduce.h"
 #include "OptimizationBackend/MatrixAccumulators.h"
 #include "vector"
 #include <math.h>
+#include <functional>
 
 namespace dso
 {
@@ -104,7 +105,7 @@ public:
 				bs[i] = VecX::Zero(nframes[0]*8+CPARS);
 			}
 
-			red->reduce(boost::bind(&AccumulatedSCHessianSSE::stitchDoubleInternal,
+			red->reduce(std::bind(&AccumulatedSCHessianSSE::stitchDoubleInternal,
 				this,Hs, bs, EF,  _1, _2, _3, _4), 0, nframes[0]*nframes[0], 0);
 
 			// sum up results
@@ -156,4 +157,3 @@ private:
 };
 
 }
-
