@@ -24,12 +24,13 @@
 
 #pragma once
 
- 
+
 #include "util/NumType.h"
 #include "OptimizationBackend/MatrixAccumulators.h"
 #include "vector"
 #include <math.h>
 #include "util/IndexThreadReduce.h"
+#include <functional>
 
 
 namespace dso
@@ -102,7 +103,7 @@ public:
 				bs[i] = VecX::Zero(nframes[0]*8+CPARS);
 			}
 
-			red->reduce(boost::bind(&AccumulatedTopHessianSSE::stitchDoubleInternal,
+			red->reduce(std::bind(&AccumulatedTopHessianSSE::stitchDoubleInternal,
 				this,Hs, bs, EF, usePrior,  _1, _2, _3, _4), 0, nframes[0]*nframes[0], 0);
 
 			// sum up results
@@ -165,4 +166,3 @@ private:
 			int min, int max, Vec10* stats, int tid);
 };
 }
-
