@@ -27,15 +27,16 @@
 
 #pragma once
 #include "util/settings.h"
-#include "boost/thread.hpp"
+#include <thread>
 #include <stdio.h>
 #include <iostream>
-
+#include <condition_variable>
+#include <functional>
 
 
 namespace dso
 {
-using namespace boost::placeholders;
+using namespace std::placeholders;
 template<typename Running>
 class IndexThreadReduce
 {
@@ -48,14 +49,14 @@ public:
 		nextIndex = 0;
 		maxIndex = 0;
 		stepSize = 1;
-		callPerIndex = boost::bind(&IndexThreadReduce::callPerIndexDefault, this, _1, _2, _3, _4);
+		callPerIndex = std::bind(&IndexThreadReduce::callPerIndexDefault, this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3, std::placeholders::_4);
 
 		running = true;
 		for(int i=0;i<NUM_THREADS;i++)
 		{
 			isDone[i] = false;
 			gotOne[i] = true;
-			workerThreads[i] = boost::thread(&IndexThreadReduce::workerLoop, this, i);
+			workerThreads[i] = std::thread(&IndexThreadReduce::workerLoop, this, i);
 		}
 
 	}
@@ -75,7 +76,7 @@ public:
 
 	}
 
-	inline void reduce(boost::function<void(int,int,Running*,int)> callPerIndex, int first, int end, int stepSize = 0)
+	inline void reduce(std::function<void(int,int,Running*,int)> callPerIndex, int first, int end, int stepSize = 0)
 	{
 
 		memset(&stats, 0, sizeof(Running));
@@ -94,7 +95,7 @@ public:
 
 		//printf("reduce called\n");
 
-		boost::unique_lock<boost::mutex> lock(exMutex);
+		std::unique_lock<std::mutex> lock(exMutex);
 
 		// save
 		this->callPerIndex = callPerIndex;
@@ -133,7 +134,7 @@ public:
 
 		nextIndex = 0;
 		maxIndex = 0;
-		this->callPerIndex = boost::bind(&IndexThreadReduce::callPerIndexDefault, this, _1, _2, _3, _4);
+		this->callPerIndex = std::bind(&IndexThreadReduce::callPerIndexDefault, this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3, std::placeholders::_4);
 
 		//printf("reduce done (all threads finished)\n");
 	}
@@ -141,13 +142,13 @@ public:
 	Running stats;
 
 private:
-	boost::thread workerThreads[NUM_THREADS];
+	std::thread workerThreads[NUM_THREADS];
 	bool isDone[NUM_THREADS];
 	bool gotOne[NUM_THREADS];
 
-	boost::mutex exMutex;
-	boost::condition_variable todo_signal;
-	boost::condition_variable done_signal;
+	std::mutex exMutex;
+	std::condition_variable todo_signal;
+	std::condition_variable done_signal;
 
 	int nextIndex;
 	int maxIndex;
@@ -155,7 +156,7 @@ private:
 
 	bool running;
 
-	boost::function<void(int,int,Running*,int)> callPerIndex;
+	std::function<void(int,int,Running*,int)> callPerIndex;
 
 	void callPerIndexDefault(int i, int j,Running* k, int tid)
 	{
@@ -165,7 +166,7 @@ private:
 
 	void workerLoop(int idx)
 	{
-		boost::unique_lock<boost::mutex> lock(exMutex);
+		std::unique_lock<std::mutex> lock(exMutex);
 
 		while(running)
 		{
