@@ -1,6 +1,6 @@
 /**
 * This file is part of DSO.
-* 
+*
 * Copyright 2016 Technical University of Munich and Intel.
 * Developed by Jakob Engel <engelj at in dot tum dot de>,
 * for more information see <http://vision.in.tum.de/dso>.
@@ -23,7 +23,7 @@
 
 
 #pragma once
-#include "boost/thread.hpp"
+#include <thread>
 #include "util/MinimalImage.h"
 #include "IOWrapper/Output3DWrapper.h"
 
