@@ -39,13 +39,13 @@ public:
     inline GTData()
     {}
 
-    inline GTData(Sophus::SE3 pose, Eigen::Vector3d velocity, Eigen::Vector3d biasRotation,
+    inline GTData(Sophus::SE3d pose, Eigen::Vector3d velocity, Eigen::Vector3d biasRotation,
                   Eigen::Vector3d biasTranslation)
             : pose(pose), velocity(velocity), biasRotation(biasRotation), biasTranslation(biasTranslation)
     {}
 
 
-    Sophus::SE3 pose;
+    Sophus::SE3d pose;
     Eigen::Vector3d velocity; // Note: velocities might be in the vicon frame instead of the world frame...
     Eigen::Vector3d biasRotation;
     Eigen::Vector3d biasTranslation;
